backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow-cpu
numpy
opencv-python-headless
Pillow
fpdf2
scikit-image